Around update hook to generate activity for updates
This uses around so we can build a meta column from the pending updates.

We perform the 'next' method before generating the activity to ensure
it's actually successful.

We can also ignore updates to columns we are not interested in
generating activity for, e.g. created_date is updated when an IA is
created, but should not change again.