Work around race condition in IA last updated tracker
If lots of updates are occurring at once, the potential for the token
to change between selecting and updating it exists. This will cause an
exception and your process will bail.

If the table has been updated in this time, we don't need to update it
since we don't need (or have) sub-second resolution in this time stamp.

This change wraps the row update in an eval - we don't care about the
result of this operation in the case where this fatal race condition
occurs.